Seperated Comment code for acceptance test case
